+++ CHANGELOG.md
@@ -5,7 +5,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
-- Prayer times with lat lon
+- Prayer times with lat lon (integrated Timer)
 - Map showing mosque locations
 - Custom Search Results
 - Bg Audio